+++ pyfont.h
@@ -2,11 +2,13 @@
 #define PYFONT_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 struct PyFont
 {
     PyFont(uint8_t chars, uint8_t baseChar, const uint8_t* data, const uint16_t* offsets, const uint8_t* sizes):
         chars(chars), baseChar(baseChar), data(data), offsets(offsets), sizes(sizes) {}
+
     uint8_t chars;
     uint8_t baseChar;
     const uint8_t* data;
@@ -26,6 +28,6 @@ struct PyFont
 
 
 int renderText(const PyFont& f, const char* text, uint8_t* output, int maxSize);
-int calculateRenderedLength(const PyFont& f, const char* text);
+size_t calculateRenderedLength(const PyFont& f, const char* text);
 
 #endif //PYFONT_H
